network/test split up into network/test/cohort1, network/test/cohort2
