+++ .github/workflows/ci.yml
@@ -100,7 +100,7 @@ jobs:
           cache: true
       - name: Set Test Matrix
         id: set-test-matrix
-        run: go run utils/test_matrix/test_matrix.go admin cmd consensus engine/access engine/collection engine/common engine/consensus engine/execution/ingestion:buildjet-8vcpu-ubuntu-2204 engine/execution/computation engine/execution engine/verification fvm ledger module network/test:buildjet-8vcpu-ubuntu-2204 network/p2p:buildjet-8vcpu-ubuntu-2204 network utils
+        run: go run utils/test_matrix/test_matrix.go admin cmd consensus engine/access engine/collection engine/common engine/consensus engine/execution/ingestion:buildjet-8vcpu-ubuntu-2204 engine/execution/computation engine/execution engine/verification fvm ledger module network/test/cohort1 network/test/cohort2 network/p2p:buildjet-8vcpu-ubuntu-2204 network utils
 
   unit-test:
     name: Unit Tests (${{ matrix.targets.name }})
