+++ network/test/cohort1/meshengine_test.go
@@ -1,4 +1,4 @@
-package test
+package cohort1
 
 import (
 	"context"
