+++ network/test/cohort2/blob_service_test.go
@@ -1,4 +1,4 @@
-package test
+package cohort2
 
 import (
 	"context"
