+++ network/test/cohort2/unicast_authorization_test.go
@@ -1,4 +1,4 @@
-package test
+package cohort2
 
 import (
 	"context"
